Comply with the linter
The E722 linter rule was violated due to a blind except in the test suite.

I have refactored the offending test in tests\powerful\test_prime.py:test_prime_creation_failure
It now uses a more conventional way in pytest to check that a line of code raises an error using `with pytest.raises`.